Ai backend/requirements.txt
fastapi>=0.68.0,<0.69.0
uvicorn>=0.15.0,<0.16.0
langchain>=0.0.350
langchain-openai>=0.0.2
langchain-community>=0.0.10
python-dotenv>=0.19.0
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.1
pydantic>=1.10.0,<2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4